requirements.txt
aiohttp==3.11.11
python-dotenv==1.0.1
pydantic==2.10.5
tenacity==9.0.0
backoff==2.2.1
requests==2.31.0
discord-webhook==1.3.0
tweepy==4.14.0
textblob==0.17.1
aiosqlite==0.19.0